pyproject.toml
[project]
name = "github-mcp-agentcore"
version = "0.1.0"
description = "GitHub MCP server adapter for Amazon Bedrock AgentCore"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore",
    "bedrock-agentcore-starter-toolkit",
    "fastmcp",
    "uvicorn",
    "aws-opentelemetry-distro",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"